+++ client/src/main.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import Host from './pages/Host';
 import Upload from './pages/Upload';
+import Booth from './pages/Booth';
 import Slideshow from './pages/Slideshow';
 import Admin from './pages/Admin';
 import './styles/global.css';
@@ -10,6 +11,8 @@ function App() {
   switch (window.location.pathname) {
     case '/upload':
       return <Upload />;
+    case '/booth':
+      return <Booth />;
     case '/slideshow':
       return <Slideshow />;
     case '/admin':